Initial attempt at batch writing instead of database usage
Program now writes derivatives in batches to the output file as it goes instead of storing them all in a database until the end, in an attempt to use less of a memory footprint. Ignores duplicates with a hash set now, which in theory should be much less memory intensive.